Refactor: moved User class package
