+++ src/main/java/com/zerobase/hseungho/restaurantreservation/global/security/SecurityHolder.java
@@ -1,6 +1,6 @@
 package com.zerobase.hseungho.restaurantreservation.global.security;
 
-import com.zerobase.hseungho.restaurantreservation.service.domain.User;
+import com.zerobase.hseungho.restaurantreservation.service.domain.user.User;
 import org.springframework.security.core.context.SecurityContextHolder;
 
 public class SecurityHolder {
