+++ src/main/java/com/zerobase/hseungho/restaurantreservation/service/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.zerobase.hseungho.restaurantreservation.service.repository;
 
-import com.zerobase.hseungho.restaurantreservation.service.domain.User;
+import com.zerobase.hseungho.restaurantreservation.service.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
