+++ src/test/java/com/zerobase/hseungho/restaurantreservation/service/appservice/UserServiceImplSignUpUnitTest.java
@@ -2,7 +2,7 @@ package com.zerobase.hseungho.restaurantreservation.service.appservice;
 
 import com.zerobase.hseungho.restaurantreservation.global.exception.impl.BadRequestException;
 import com.zerobase.hseungho.restaurantreservation.global.exception.model.ErrorCodeType;
-import com.zerobase.hseungho.restaurantreservation.service.domain.User;
+import com.zerobase.hseungho.restaurantreservation.service.domain.user.User;
 import com.zerobase.hseungho.restaurantreservation.service.dto.SignUp;
 import com.zerobase.hseungho.restaurantreservation.service.dto.UserDto;
 import com.zerobase.hseungho.restaurantreservation.service.repository.UserRepository;
